pyproject.toml
[project]
name = "pygame1"
version = "0.1.0"
description = "A simple 2D physics simulation using Pymunk and Pygame"
dependencies = [
    "pygame-ce",
    "pymunk>=7.1.0",
]
requires-python = ">=3.8"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "black>=24.4.2",
    "pytest>=8.3.2",
]
